requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.6b0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
